moves shared calculator functions to solids manure calc class
