+++ RUFAS/biophysical/manure/storage/composting_type.py
@@ -1,0 +1,21 @@
+from enum import Enum
+
+
+class CompostingType(Enum):
+    """
+    This is an Enum class that represents different types of composting.
+
+    Attribute
+    ----------
+    INTENSIVE_WINDROW : str
+        Intensive windrow.
+    PASSIVE_WINDROW : str
+        Passive windrow.
+    STATIC_PILE: str
+        Static pile.
+
+    """
+
+    INTENSIVE_WINDROW = "intensive windrow"
+    PASSIVE_WINDROW = "passive windrow"
+    STATIC_PILE = "static pile"
