+++ tests/test_biophysical/test_manure/test_storage/test_composting.py
@@ -3,55 +3,21 @@ import math
 import pytest
 from pytest_mock import MockerFixture
 from RUFAS.biophysical.manure.storage.composting import (
-    AMBIENT_AIR_MOLE_FRACTION_OF_OXYGEN,
-    DEFAULT_CARBON_FRACTION_AVAILABLE_IN_VSND,
-    DEFAULT_CARBON_FRACTION_AVAILABLE_IN_VSD,
-    COMPOSTING_DECOMPOSITION_TEMPERATURE,
-    DEFAULT_DAYS_SINCE_LAST_TURNING,
-    DEFAULT_EFFECT_OF_MOISTURE_ON_MICROBIAL_DECOMPOSITION,
-    EFFECTIVENESS_OF_MICROBIAL_DECOMPOSITION_RATE,
-    FIRST_ORDER_DECAYING_COEFFICIENT,
-    DEFAULT_LAG_TIME,
-    DEFAULT_MOLE_FRACTION_OF_OXYGEN,
     MCF_COMPOSTING_STATIC_PILE,
     MCF_COMPOSTING_WINDROW_HIGH,
     MCF_COMPOSTING_WINDROW_LOW,
     MCF_COMPOSTING_WINDROW_MEDIUM,
     MCF_LOWER_BOUND_TEMPERATURE,
     MCF_UPPER_BOUND_TEMPERATURE,
-    OXYGEN_HALF_SATURATION_CONSTANT,
     Composting,
-    CompostingType,
 )
+from RUFAS.biophysical.manure.storage.composting_type import CompostingType
+from RUFAS.biophysical.manure.storage.solids_storage_calculator import SolidsStorageCalculator
 from RUFAS.biophysical.manure.storage.storage_cover import StorageCover
 from RUFAS.data_structures.animal_to_manure_connection import ManureStream
 from RUFAS.current_day_conditions import CurrentDayConditions
 from RUFAS.output_manager import OutputManager
 from RUFAS.rufas_time import RufasTime
-
-
-def test_enum_members_exist() -> None:
-    assert CompostingType.INTENSIVE_WINDROW.name == "INTENSIVE_WINDROW"
-    assert CompostingType.PASSIVE_WINDROW.name == "PASSIVE_WINDROW"
-    assert CompostingType.STATIC_PILE.name == "STATIC_PILE"
-
-
-def test_enum_values() -> None:
-    assert CompostingType.INTENSIVE_WINDROW.value == "intensive windrow"
-    assert CompostingType.PASSIVE_WINDROW.value == "passive windrow"
-    assert CompostingType.STATIC_PILE.value == "static pile"
-
-
-def test_enum_reverse_lookup() -> None:
-    assert CompostingType("intensive windrow") == CompostingType.INTENSIVE_WINDROW
-    assert CompostingType("passive windrow") == CompostingType.PASSIVE_WINDROW
-    assert CompostingType("static pile") == CompostingType.STATIC_PILE
-
-
-def test_invalid_enum_raises_value_error() -> None:
-    with pytest.raises(ValueError) as exc_info:
-        CompostingType("deep pit composting")
-    assert "deep pit composting" in str(exc_info.value)
 
 
 def test_composting_init(mocker: MockerFixture) -> None:
@@ -131,12 +97,12 @@ def test_process_manure_runs_expected_steps(
         composting_instance, "_calculate_composting_methane_emissions", return_value=1.0
     )
     mock_calc_carb_decomp = mocker.patch.object(
-        composting_instance, "_calculate_carbon_decomposition", return_value=1.0
+        SolidsStorageCalculator, "_calculate_carbon_decomposition", return_value=1.0
     )
     mock_apply_dml = mocker.patch.object(composting_instance, "_apply_dry_matter_loss")
     mock_calc_n2o = mocker.patch.object(composting_instance, "_calculate_nitrous_oxide_emissions", return_value=0.5)
     mock_calc_leaching = mocker.patch.object(
-        composting_instance, "_calculate_nitrogen_loss_to_leaching", return_value=0.5
+        SolidsStorageCalculator, "_calculate_nitrogen_loss_to_leaching", return_value=0.5
     )
     mock_calc_ammonia = mocker.patch.object(
         composting_instance, "_calculate_composting_ammonia_emissions", return_value=0.5
@@ -186,7 +152,7 @@ def test_apply_dry_matter_loss_valid(
     composting_instance._received_manure = received_manure
     composting_instance._manure_to_process = copy(received_manure)
     mocker.patch.object(
-        composting_instance,
+        SolidsStorageCalculator,
         "_calculate_dry_matter_loss",
         return_value=4.0,
     )
@@ -217,7 +183,7 @@ def test_apply_dry_matter_loss_raises_value_error(
     composting_instance._om = OutputManager()
     mock_add_error = mocker.patch.object(composting_instance._om, "add_error", return_value=None)
     mocker.patch.object(
-        composting_instance,
+        SolidsStorageCalculator,
         "_calculate_dry_matter_loss",
         return_value=100.0,
     )
@@ -300,17 +266,6 @@ def test_calculate_composting_ammonia_emissions() -> None:
     assert result == pytest.approx(expected)
 
 
-def test_calculate_nitrogen_loss_to_leaching() -> None:
-    """Test nitrogen loss to leaching calculation with a simple input."""
-    composting_type = CompostingType.PASSIVE_WINDROW
-    received_nitrogen = 20.0
-
-    expected = 0.04 * 20.0
-    result = Composting._calculate_nitrogen_loss_to_leaching(composting_type, received_nitrogen)
-
-    assert result == pytest.approx(expected)
-
-
 def test_calculate_nitrous_oxide_emissions() -> None:
     """Test nitrous oxide emission calculation with a simple input."""
     nitrous_oxide_fraction = 0.02
@@ -360,84 +315,3 @@ def test_calculate_methane_conversion_factor(
     """Test MCF value returned based on composting type and temperature."""
     result = Composting._calculate_methane_conversion_factor(temperature, composting_type)
     assert result == expected
-
-
-def test_calculate_dry_matter_loss() -> None:
-    """Test dry matter loss calculation."""
-    methane_emissions = 0.12
-    carbon_decomposition = 0.24
-
-    expected = 2 * carbon_decomposition + methane_emissions
-    result = Composting._calculate_dry_matter_loss(methane_emissions, carbon_decomposition)
-
-    assert result == pytest.approx(expected)
-
-
-def test_calculate_carbon_decomposition(mocker: MockerFixture) -> None:
-    """Test carbon decomposition calculation with mocked coefficient + rate."""
-    manure_temp = 30.0
-    total_solids = 10.0
-    ndvs = 5.0
-
-    mocker.patch.object(Composting, "_calculate_carbon_decomposition_rate", return_value=0.1)
-    mocker.patch.object(Composting, "_calculate_anaerobic_coefficient", return_value=0.2)
-
-    expected = (
-        (total_solids * DEFAULT_CARBON_FRACTION_AVAILABLE_IN_VSD + ndvs * DEFAULT_CARBON_FRACTION_AVAILABLE_IN_VSND)
-        * 0.1
-        * DEFAULT_EFFECT_OF_MOISTURE_ON_MICROBIAL_DECOMPOSITION
-        * 0.2
-    )
-
-    result = Composting._calculate_carbon_decomposition(manure_temp, ndvs, total_solids)
-    assert result == pytest.approx(expected)
-
-
-def test_calculate_carbon_decomposition_rate(mocker: MockerFixture) -> None:
-    """Test carbon decomposition rate with mocked decomposition values."""
-    manure_temp = 30.0
-    r_max = 0.2
-    r_slow = 0.05
-
-    mocker.patch.object(Composting, "_calculate_max_microbial_decomposition_rate", return_value=r_max)
-    mocker.patch.object(Composting, "_calculate_slow_fraction_decomposition_rate", return_value=r_slow)
-
-    exponent = FIRST_ORDER_DECAYING_COEFFICIENT * (DEFAULT_DAYS_SINCE_LAST_TURNING - DEFAULT_LAG_TIME)
-    expected = (r_max - r_slow) * (math.e**exponent) + r_slow
-
-    result = Composting._calculate_carbon_decomposition_rate(manure_temp)
-    assert result == pytest.approx(expected)
-
-
-def test_calculate_max_microbial_decomposition_rate() -> None:
-    """Test that the max microbial decomposition rate is computed correctly."""
-    expected = float(
-        EFFECTIVENESS_OF_MICROBIAL_DECOMPOSITION_RATE
-        * (1.066 ** (COMPOSTING_DECOMPOSITION_TEMPERATURE - 10) - 1.21 ** (COMPOSTING_DECOMPOSITION_TEMPERATURE - 50))
-    )
-
-    result = Composting._calculate_max_microbial_decomposition_rate()
-    assert result == pytest.approx(expected)
-
-
-def test_calculate_slow_fraction_decomposition_rate() -> None:
-    """Test slow fraction decomposition rate calculation at a specific temperature."""
-    manure_temperature = 35.0
-
-    expected = float(
-        EFFECTIVENESS_OF_MICROBIAL_DECOMPOSITION_RATE
-        * (1.066 ** (manure_temperature - 10) - 1.21 ** (manure_temperature - 50))
-    )
-
-    result = Composting._calculate_slow_fraction_decomposition_rate(manure_temperature)
-    assert result == pytest.approx(expected)
-
-
-def test_calculate_anaerobic_coefficient() -> None:
-    """Test anaerobic coefficient calculation against expected value."""
-    expected = (
-        DEFAULT_MOLE_FRACTION_OF_OXYGEN / (OXYGEN_HALF_SATURATION_CONSTANT + DEFAULT_MOLE_FRACTION_OF_OXYGEN)
-    ) * ((OXYGEN_HALF_SATURATION_CONSTANT + AMBIENT_AIR_MOLE_FRACTION_OF_OXYGEN) / AMBIENT_AIR_MOLE_FRACTION_OF_OXYGEN)
-
-    result = Composting._calculate_anaerobic_coefficient()
-    assert result == pytest.approx(expected)
